Fix: Export jdk modules for spotlessApply
