+++ src/main/java/de/sean/blockprot/bukkit/inventories/BlockProtInventory.java
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package de.sean.blockprot.bukkit.inventories;
 
 import de.sean.blockprot.TranslationKey;
