+++ src/main/java/de/sean/blockprot/util/SemanticVersion.java
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package de.sean.blockprot.util;
 
 import org.jetbrains.annotations.NotNull;
