+++ src/main/kotlin/de/sean/blockprot/bukkit/inventories/BlockInfoInventory.kt
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package de.sean.blockprot.bukkit.inventories
 
 import de.sean.blockprot.BlockProt
