+++ src/main/kotlin/de/sean/blockprot/bukkit/nbt/FriendModifyAction.kt
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package de.sean.blockprot.bukkit.nbt
 
 enum class FriendModifyAction {
